split groups and abstract_groups
