+++ backend/apps/academic/templates/academic/detail.html
@@ -1,4 +1,4 @@
-{% extends 'group/detail/detail.html' %}
+{% extends 'abstract_group/detail/detail.html' %}
 
 {% block object_name %}
     <h1>{{object.name}}</h1>
