+++ backend/apps/club/templates/club/detail.html
@@ -1,4 +1,4 @@
-{% extends 'group/detail/detail.html' %}
+{% extends 'abstract_group/detail/detail.html' %}
 
 {% block infos %}
     {% if object.email or object.meeting_place or object.meeting_hour %}
