+++ backend/apps/family/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.urls.base import reverse
 from django.utils import timezone
 
-from apps.group.models import AbstractGroup, NamedMembership
+from apps.group.abstract.models import AbstractGroup, NamedMembership
 from apps.student.models import Student
