+++ backend/apps/group/templates/abstract_group/edit/event/archived_edit.html
@@ -3,7 +3,7 @@
 {% load static %}
 
 {% block content %}
-{% include 'group/included/navbar_edit.html' with open='events' %}
+{% include 'abstract_group/included/navbar_edit.html' with open='events' %}
 
 <h2>Liste des events archivés</h2>
 <div id="root"></div>
