+++ backend/apps/group/templates/abstract_group/edit/event/create.html
@@ -2,7 +2,7 @@
 {% load crispy_forms_tags %}
 
 {% block content %}
-{% include 'group/included/navbar_edit.html' with open='events' %}
+{% include 'abstract_group/included/navbar_edit.html' with open='events' %}
 
 <form method='POST' enctype="multipart/form-data">
   {{ form.media }}
