+++ backend/apps/group/templates/abstract_group/edit/members_edit.html
@@ -3,7 +3,7 @@
 {% load static %}
 
 {% block content %}
-{% include 'group/included/navbar_edit.html' with open='members' %}
+{% include 'abstract_group/included/navbar_edit.html' with open='members' %}
 
 <h2>Membres</h2>
 <!--<form method='POST' action="{% url object.app|add:':update-members' object.slug %}">
