+++ backend/apps/group/templates/abstract_group/edit/post/all_edit.html
@@ -3,7 +3,7 @@
 
 {% block content %}
 
-{% include 'group/included/navbar_edit.html' with open='posts' %}
+{% include 'abstract_group/included/navbar_edit.html' with open='posts' %}
 
 <h2>Liste des annonces</h2>
 {% if posts.count > 0 %}
