+++ backend/apps/group/templates/abstract_group/edit/post/create.html
@@ -3,7 +3,7 @@
 
 {% block content %}
 
-  {% include 'group/included/navbar_edit.html' with open='posts' %}
+  {% include 'abstract_group/included/navbar_edit.html' with open='posts' %}
 
   <form method='POST' enctype="multipart/form-data">
     {{ form.media }}
