+++ backend/apps/group/templates/abstract_group/edit/sociallinks_edit.html
@@ -2,7 +2,7 @@
 {% load crispy_forms_tags %}
 
 {% block content %}
-{% include 'group/included/navbar_edit.html' with open='sociallinks' %}
+{% include 'abstract_group/included/navbar_edit.html' with open='sociallinks' %}
 
 <h2>Liens de Réseaux Sociaux</h2>
 <form method='POST' action="{% url object.app|add:':update-sociallinks' object.slug %}">
