+++ backend/apps/group/templates/abstract_group/edit/update.html
@@ -2,7 +2,7 @@
 {% load crispy_forms_tags %}
 
 {% block content %}
-{% include 'group/included/navbar_edit.html' with open='general' %}
+{% include 'abstract_group/included/navbar_edit.html' with open='general' %}
 
 <form method='POST' enctype="multipart/form-data">
     {{ form.media }}
