+++ backend/apps/group/templates/abstract_group/mail/remove_admin.html
