+++ backend/apps/post/templates/post/update.html
@@ -2,7 +2,7 @@
 {% load crispy_forms_tags %}
 {% block content %}
 
-{% include 'group/included/navbar_edit.html' with open='posts' %}
+{% include 'abstract_group/included/navbar_edit.html' with open='posts' %}
 
 <form method='POST' enctype="multipart/form-data">
   {% csrf_token %}
