+++ backend/apps/roommates/templates/roommates/coloc/detail/detail.html
@@ -1,4 +1,4 @@
-{% extends 'group/detail/detail.html' %}
+{% extends 'abstract_group/detail/detail.html' %}
 {% load static %}
 {% load file_url %}
