+++ backend/apps/roommates/urls.py
@@ -8,7 +8,7 @@ from .views import (
     ColocathlonFormView,
     DetailRoommatesView,
     UpdateRoommatesView)
-from apps.group.urls import make_group_url_patterns
+from apps.group.abstract.urls import make_group_url_patterns
 
 app_name = 'roommates'
