+++ backend/apps/utils/upload.py
@@ -11,7 +11,7 @@ class PathAndRename(object):
 
     def __call__(self, instance, filename):
         ext = filename.split('.')[-1]
-        from apps.group.models import AbstractGroup
+        from apps.group.abstract.models import AbstractGroup
         if isinstance(instance, AbstractGroup):
             dir = os.path.join(self.path, instance.app)
         else:
